fix(admin): room - align with review feedback
- auditClient: default to slog.Default() when a nil logger is passed;
  warn when only one of INTERNAL_CLIENT_CERT/INTERNAL_CLIENT_KEY is set;
  log CA file read/parse failures instead of ignoring them
- CountRooms: scan COUNT(*) into int64 (matches Postgres COUNT(*) width)
  and return int64, dropping the service-layer cast; interface/mocks/tests
  updated accordingly
- tests: audit client nil-logger and partial-cert-config coverage

+++ backend/services/room/handler/adminRoomHandler_test.go
@@ -239,7 +239,7 @@ func TestAdminRoomHandler_StatsRooms_Success(t *testing.T) {
 	roomRepo := new(adminRoomRepoMock)
 	h := newAdminRoomHandler(t, roomRepo, new(adminMemberRepoMock))
 
-	roomRepo.On("CountRooms", mock.Anything).Return(12, nil).Once()
+	roomRepo.On("CountRooms", mock.Anything).Return(int64(12), nil).Once()
 
 	c, rec := echotest.ContextConfig{}.ToContextRecorder(t)
 	setAdminRoomClaims(c)
